Move expand to a module
